feat: merge the two experiment panels
