+++ packages/visualisation/src/components/SideMenu/index.js
@@ -1,7 +1,6 @@
 import React, { useState, useRef } from 'react'
 import styled from 'styled-components'
 import pmLogo from '../../icons/svg/pmLogo.svg'
-import plus from '../../icons/svg/plus.svg'
 import experimentIcon from '../../icons/svg/experimentIcon.svg'
 import historyIcon from '../../icons/svg/historyIcon.svg'
 import menuActive from '../../icons/svg/menuActive.svg'
@@ -9,7 +8,6 @@ import pencilIcon from '../../icons/svg/pencil.svg'
 import SavedExperimentSection from '../SavedExperimentSection'
 import ExperimentSection from '../ExperimentSection'
 import NewExperimentSection from '../NewExperimentSection'
-import ExperimentParametersSection from '../ExperimentParametersSection'
 import useOutsideClick from '../../hooks/useClickOutside'
 import { keyframes } from 'styled-components'
 
@@ -98,24 +96,8 @@ const SideMenu = ({
               )
             }
           >
-            <img src={plus} alt="New Experiment" />
-            {open === 'newExperiment' && (
-              <ActiveMenu>
-                <img src={menuActive} alt="Open" />
-              </ActiveMenu>
-            )}
-          </MenuItem>
-          <MenuItem
-            onClick={() =>
-              setOpen((current) =>
-                current === 'experimentParameters'
-                  ? 'map'
-                  : 'experimentParameters'
-              )
-            }
-          >
             <img src={pencilIcon} alt="Parametrar" />
-            {open === 'experimentParameters' && (
+            {open === 'newExperiment' && (
               <ActiveMenu>
                 <img src={menuActive} alt="Open" />
               </ActiveMenu>
@@ -160,17 +142,10 @@ const SideMenu = ({
       )}
 
       {open === 'savedExperiment' && <SavedExperimentSection />}
-      {open === 'experimentParameters' && (
-        <ExperimentParametersSection
-          fleets={fleets}
-          newExperiment={newExperiment}
-          newParameters={newParameters}
-          setNewParameters={setNewParameters}
-        />
-      )}
 
       {open === 'newExperiment' && (
         <NewExperimentSection
+          fleets={fleets}
           newParameters={newParameters}
           newExperiment={newExperiment}
           setNewParameters={setNewParameters}
